Fix save/unsave recipe controller

+++ Development/NutriMon/src/main/java/chillingMonsters/Pages/recipePage/RecipeCardComponent.java
@@ -26,7 +26,7 @@ public class RecipeCardComponent extends AnchorPane {
 	@FXML
 	private Label cardCalories;
 
-	public RecipeCardComponent(long recipeID, String name, String category, float time, float calories) {
+	public RecipeCardComponent(long recipeID, String name, String category, int time, float calories) {
 		super();
 		this.recipeID = recipeID;
 
@@ -49,7 +49,7 @@ public class RecipeCardComponent extends AnchorPane {
 
 		cardName.setText(name);
 		cardCategory.setText(category.toUpperCase());
-		cardCookTime.setText(String.format("%.0f mins", time));
+		cardCookTime.setText(String.format("%d mins", time));
 		cardCalories.setText(String.format("%.0f kcal", calories));
 	}
